GroundingDINO_space/requirements.txt
fastapi
gradio
torch
Pillow
numpy
groundingdino
